pyproject.toml
[tool.poetry]
name = "django-sortedm2m-filter-horizontal-widget"
version = "2.0.1"
description = "Django admin widget for Gregor Mülleggers django-sortedm2m_ library"
authors = ["Sander van Leeuwen <replytosander@gmail.com>"]
license = "BSD"
readme = "README.md"
packages = [{include = "sortedm2m_filter_horizontal_widget"}]

[tool.poetry.dependencies]
python = "^3.8"
django-sortedm2m = "^3.1.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
